refactor: relocate appbar to components folder
